fix all 12 clippy warnings for clean CI
- empty_line_after_doc_comment: context_pack.rs
- manual_clamp: dag.rs (use .clamp())
- map_entry: dag.rs (use Entry::Vacant)
- type_complexity: db.rs (3 functions)
- needless_borrow: runtime.rs (2 occurrences)
- redundant_closure: replay.rs
- unnecessary_cast: summarizer.rs (2 occurrences)
- private_interfaces: dag.rs (DagGraph -> pub(crate))

+++ src/summarizer.rs
@@ -480,7 +480,7 @@ impl Summarizer {
                         }
                     } else if status.as_u16() == 429 {
                         // Rate limited — full-jitter exponential backoff
-                        let delay_ms = crate::runtime::full_jitter_backoff(2000, 60000, attempt as u32, jitter_seed());
+                        let delay_ms = crate::runtime::full_jitter_backoff(2000, 60000, attempt, jitter_seed());
                         let delay = Duration::from_millis(delay_ms);
                         tracing::warn!(
                             target = "deeplossless::summarizer",
@@ -516,7 +516,7 @@ impl Summarizer {
                         "request failed"
                     );
                     if e.is_timeout() || e.is_connect() {
-                        let delay_ms = crate::runtime::full_jitter_backoff(1000, 30000, attempt as u32, jitter_seed());
+                        let delay_ms = crate::runtime::full_jitter_backoff(1000, 30000, attempt, jitter_seed());
                         let delay = Duration::from_millis(delay_ms);
                         tokio::time::sleep(delay).await;
                         last_error = Some(anyhow::anyhow!("{e}"));
